fix(stella-tui): unbreak main — deck ratchet, StepUsage fixtures, stale cache tests
Three co-landing breakages on main, none introduced here:

1. File-size ratchet: deck.rs 1804 > 1776 and deck_render.rs 2107 > 2055
   after #300's cache-panel work. Extracted each file's inline
   `#[cfg(test)] mod tests` to a sibling `deck/tests.rs` /
   `deck_render/tests.rs` (the crate's established pattern) and LOWERED
   both recorded counts — never raised. deck.rs 1804->1022,
   deck_render.rs 2107->1470.

2. `AgentEvent::StepUsage` gained `role`, `provider` and `complete`, but
   three fixtures still used the old shape, so `cargo test -p stella-tui`
   did not compile: cache_panel.rs, deck_render/tests.rs, views/agents.rs.
   Added the missing fields.

3. Two stale cache-statline tests in deck_render/tests.rs asserted an
   obsolete format `(105.3M/211.4M tokens)` that the shipped renderer no
   longer produces — it renders `(105.3M rd · 0 wr)`, which is what #267
   asks for (read AND write volumes). cache_panel.rs already carries the
   equivalent assertions in the current format, so the duplicates were
   removed rather than the renderer changed. No coverage lost.

Verified: cargo test -p stella-tui 497 passed / 0 failed, clippy
--all-targets -D warnings clean, cargo fmt --all --check clean,
check-file-sizes OK (340 files).
