+++ stella-tui/src/cache_panel.rs
@@ -336,6 +336,8 @@ mod tests {
     fn step_usage_full(input: u64, cached: u64, write: u64) -> AgentEvent {
         AgentEvent::StepUsage {
             step: 1,
+            role: stella_protocol::event::ModelCallRole::Worker,
+            provider: "zai".into(),
             model: "glm".into(),
             input_tokens: input,
             output_tokens: 0,
@@ -346,6 +348,7 @@ mod tests {
             duration_ms: 1,
             retries: 0,
             tool_calls: 0,
+            complete: true,
         }
     }
